api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0